Changes to Improve Segmentation
Changes to splitFluo.m to make intermediate images easier to check.
Reduced the texture threshold in splitBFseries.m
Added an open(3), close(6) to the output of the brightfield segmentation to eliminate small specks.
Added saving of brightfield and individual fluorescent channel segmentations to analyseImageContactData and batch version
minor changes to output names,