Timelapse velocity image analysis scripts
Timelapse_bg_subtract.ijm is run on tif stacks and generates a folder with individual images for each timepoint, for each experiment. Images are also divided by a blurred median as an illumination correction.
PIVlab_commandline_batch_process.m is then run, taking each experiment's folder (which contains the separated timepoint folders) as an input array. This script is based on the PIVlab batch script and runs the PIV analysis on each timepoint, producing a .m result file.
PIVlabDataExtract_V3.m is then run, again taking each experiment's folder (which contains the separated timepoint folders) as an input array. This script reads the PIVlab data and outputs the measurements we want in the format we want and saves it as a .csv.
PIV_analyze_V2.R is then run. It automatically finds the .csv files based on a shared naming convention with the last matlab script. It also takes as input a separate folder with the fluorescence blocks which have been processed and organized. It then loops through each experiment, loading the PIV data and the fluorescent mask for refining the PIV measurements to locations where fluorescence was detected. It then calculates the mean, sd, median and log transformed equivalents of velocity magnitude to summarize the velocity measurements across space and time for each timepoint. It also makes some plots to show how log velocities are normally distributed. It outputs a single csv that contains the velocity measurements for all timepoints for all experiments.